requirements.txt
Flask>1.1.4,<3
werkzeug>2.0
flask-session>=0.4
python-dotenv
SQLAlchemy==1.4.39
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
psycopg2-binary
requests
itsdangerous==2.0.1
Jinja2==3.1.2



flake8==5.0.4
isort==4.3.21
pipenv==2022.1.8
python-dateutil==2.8.2
urllib3>=1.26.14
